backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
ultralytics==8.0.200
torch==2.5.1
torchvision==0.20.1
websockets==12.0
numpy==1.26.2
pillow==10.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
